+++ src/Json/JsonValue.fs
@@ -216,6 +216,24 @@ type private JsonParser(jsonText:string, cultureInfo, tolerateErrors) =
                     let ch = unicodeChar (s.Substring(i+2, 4))
                     buf.Append(ch) |> ignore
                     i <- i + 4  // the \ and u will also be skipped past further below
+                | 'U' ->
+                    ensure(i+9 < s.Length)
+                    let unicodeChar (s:string) =
+                        if s.Length <> 8 then failwith "unicodeChar";
+                        if s.[0..1] <> "00" then failwith "unicodeChar";     // only code points U+010000 to U+10FFFF supported
+                                                                             // for coversion to UTF16 surrogate pairs
+                        // used http://en.wikipedia.org/wiki/UTF-16#Code_points_U.2B010000_to_U.2B10FFFF as a guide below
+                        let codePoint = System.UInt32.Parse(s, NumberStyles.HexNumber) - 0x010000u
+                        let HIGH_TEN_BIT_MASK = 0xFFC00u                     // 1111|1111|1100|0000|0000
+                        let LOW_TEN_BIT_MASK = 0x003FFu                      // 0000|0000|0011|1111|1111
+                        let leadSurrogate = (codePoint &&& HIGH_TEN_BIT_MASK >>> 10) + 0xD800u
+                        let trailSurrogate = (codePoint &&& LOW_TEN_BIT_MASK) + 0xDC00u                        
+                        char leadSurrogate, char trailSurrogate
+
+                    let lead, trail = unicodeChar (s.Substring(i+2, 8))
+                    buf.Append(lead) |> ignore
+                    buf.Append(trail) |> ignore
+                    i <- i + 8  // the \ and u will also be skipped past further below
                 | _ -> throw()
                 i <- i + 2  // skip past \ and next char
             else